2.1.0: executor: Refactored. Removed thread priorities to simplify API. git-svn-id: https://casparcg.svn.sourceforge.net/svnroot/casparcg/server/branches/2.1.0@2233 362d55ac-95cf-4e76-9f9a-cbaa9c17b72d

+++ common/prec_timer.h
@@ -33,7 +33,7 @@ public:
 	void tick(double interval);
 
 private:	
-	DWORD time_;
+	unsigned long time_;
 };
